requirements.txt
flask>=3.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.3.0
snowleopard>=0.2.0
python-dotenv>=1.0.0